% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf04.r
\name{adf04}
\alias{adf04}
\title{ADF test for PANIC (2004)}
\usage{
adf04(y,k,p)
}
\arguments{
\item{y}{A matrix containing the data.}

\item{k}{An integer specifying the maximum lag order for individual
ADF regressions. Bai and Ng (2004) suggest 4*(Time/100)^(.25) rounded
to the nearest whole number as the maximum number of lags.}

\item{p}{An integer of value either 0 or 1. p is the order of the deterministic
function in the regression. 0 is for constant only and 1 is to include a trend.}
}
\value{
tstat A vector of t statistics for each column of the input matrix
}
\description{
This function performs the ADF tests on the idiosyncratic and common components
for PANIC (2004).
}

