% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glsd}
\alias{glsd}
\title{General Least Squares Detrending}
\usage{
glsd(y,p)
}
\arguments{
\item{y}{The dataframe to be detreneded}

\item{p}{an binary input of 0 or 1 which decides value of penalty chat of either -7
or -13.5, respectively}
}
\value{
yt The data detrended
}
\description{
This function detrends the data by general least squares.
}

