% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lagn}
\alias{lagn}
\title{Create lags for matrix}
\usage{
lagn(x,n)
}
\arguments{
\item{x}{A 1XT vector to apply the lag to}

\item{n}{the number of lags to add to the vector}
}
\value{
y A 1XT vector with a lag n added
}
\description{
This function adds lags to a vector.
}

