% Generated by roxygen2 (4.0.2): do not edit by hand
\name{minindc}
\alias{minindc}
\title{Create an Index of lowest values of each column for a matrix}
\usage{
minindc(x)
}
\arguments{
\item{x}{A matrix that will be be used to create the index of lowest values.}
}
\value{
pos A vector containig an index of the row position of the lowest
value within each column of a matrix.
}
\description{
This function creates a vector filled with the row position
of the lowest value within each column of a matrix.
}

