% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mydiff}
\alias{mydiff}
\title{Difference a matrix}
\usage{
mydiff(x,k)
}
\arguments{
\item{x}{The NxT matrix to be differenced.}

\item{k}{The length of difference to perform on each vector of the NxT matrix.}
}
\value{
xx A matrix that has been differenced by k periods ago.
}
\description{
This function differences an entire matrix by some defined
constant k.
}

