% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myols.r
\name{myols}
\alias{myols}
\title{Beta Coefficients for standard OLS}
\usage{
myols(x,y)
}
\arguments{
\item{x}{A matrix of dependent variables.}

\item{y}{A vector that is used as the independent variable.}
}
\description{
Returns the Beta values for an Ordinary Least Squares.
}

