% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIPAagg5.R
\docType{data}
\name{NIPA_agg_5}
\alias{NIPA_agg_5}
\title{NIPA Aggregate Level 5}
\format{An xts object with 677 rows and 17 columns. All columns are numeric with titles above}
\source{
http://www.bea.gov/iTable/iTable.cfm?ReqID=12&step=1#reqid=12&step=1&isuri=1
}
\description{
A dataset containing an aggregate of the National Income and Product Account monthly data
}

