% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{NIPAagg2}
\alias{NIPAagg2}
\title{NIPA Aggregate Level 2}
\format{A data frame with 648 rows and 45 columns. All columns are numeric with titles above}
\source{
http://www.bea.gov/iTable/iTable.cfm?ReqID=12&step=1#reqid=12&step=1&isuri=1
}
\description{
A dataset containing an aggregate of the National Income and Product Account monthly data
}

