% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nw}
\alias{nw}
\title{Bandwidth selection}
\usage{
nw(v,fixk)
}
\arguments{
\item{v}{A Vector of error terms from a factor model}

\item{fixk}{If fixk is 0, then this function will perform automatic bandwidth
selection. Otherwise, the integer placed here will be the selected bandwidth.}
}
\value{
k An integer that is the bandwidth chosen by \code{nw()}

w The vector of long run variance
}
\description{
This function at fixk 0 finds the bandwidth, or lag length, of the
short, long, and run variance of a residual.
}
\references{
Moon, R. & B. Perron (2004) Testing for a unit root in panels with
dynamic factors. Journal of Econometrics 122, 81-126.
}

