% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cPANIC.r
\name{panic04}
\alias{panic04}
\title{PANIC (2004) Non-Stationarity Tests on Common and Idiosyncratic Components}
\usage{
panic04(x, nfac, k1, criteria)
}
\arguments{
\item{x}{An object of class xts holding the time series data}

\item{nfac}{An integer speciyfing the maximum number of factors allowed
while estimating the factor model.}

\item{k1}{The maximum lag allowed in the ADF test.}

\item{criteria}{a character vector with values of either IC1, IC2, IC3, AIC1, BIC1, AIC3, BIC3, or eigen.
Choosing eigen makes the number of factors equal to the number of columns whose sum of eigenvalues is less than  or equal to .5.}
}
\value{
adff A data frame containing pooled demeaned critical values, demeaned error term critical values ,demeaned
 and detrended critical values ,R squared for principle component
 , and the significance of the error components.

adf.ind A matrix containing the critical values for the pooled Demeaned ADF test, the
pooled ADF test on the common components, the pooled demeaned ADF test on the
Idiosyncratic component, and the pooled first differenced and demeaned ADF test on the
Idiosyncratic component.
}
\description{
This function performs the pooled tests on the idiosyncratic and common component from PANIC (2004).
}
\references{
Bai, Jushan, and Serena Ng. 
'A PANIC Attack on Unit Roots and Cointegration.'
Econometrica 72.4 (2004): 1127-1177. Print.
}

