% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panic10lm}
\alias{panic10lm}
\title{PANIC (2010) Demeanded tests for Idiosyncratic Component}
\usage{
panic10lm(x, nfac, p, k1, jj)
}
\arguments{
\item{x}{A NxT matrix containing the data}

\item{nfac}{An integer specifying the maximum number of factors allowed
while estimating the factor model.}

\item{p}{A binary selection for 0 or 1. p is the order of the determinisitic
function in the regression. 0 is for constant only and 1 is to include a trend.}

\item{k1}{The maximum lag allowed in the ADF test.}

\item{jj}{an Integer 1 through 8. Choices 1 through 7 are respectively, IC(1),
IC(2), IC(3), AIC(1), BIC(1), AIC(3), and BIC(3), respectively. Choosing 8
makes the number of factors equal to the number of columns whose sum of
eigenvalues is less than  or equal to .5.}
}
\value{
rho1 Estimation of the Pooled Autoregressive Coefficient.

Model This function shows MP model C. Model C assumes a constant and trend

test.C.P A matrix containing the Pooled tests as well as Model C.

extra.test a matrix containing t_c, rho1, and the LM test for PANIC (2004)'s
Idiosyncratic Components.

t_a-t_b Pooled test from PANIC (2010). Null of nonstationarity.
If both reject conclude stationarity. However, if only one rejects the panel
is nonstationary.

t_a1-t_a2 Modified Moon and Perron test with constant and trend (Model C).
Null of nonstationarity. If both reject conclude stationarity. However,
if only one rejects the panel is nonstationary.

PMSB Unit root test tends to zero. The unit root hypothesis is rejected
in favor of stationarity when the PMSB test goes below a critical value
}
\description{
This function performs PANIC (2010) Model C, PAC, and PMSB tests.
PAC estimates the pooled autoregressive coefficient, PMSB uses a sample
moment, and Model C performs the MP test while projecting on intercept and trend.
 The sample moments test is based off of the modified Sargan-Bhargava test (PMSB).
}
\references{
pa-pb The pooled tests on the idiosyncratic component
}

