% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.R
\name{pc}
\alias{pc}
\title{Principle Component Analysis for PANIC (2004)}
\usage{
pc(y,nfac)
}
\arguments{
\item{y}{An NxT matrix containing the data}

\item{nfac}{An integer specifying the maximum number of factors allowed
while estimating the factor model.}
}
\value{
ehat A matrix with the Idiosyncratic component of the factor model

fhat A matrix with the factors of the approximate factor model

lambda A matrix with the factor loadings of the approximate factor model
}
\description{
This function performs the principle component analysis in order to determine
the Common and Idiosyncratic Components of the factor model.
}

