% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool10.r
\name{pool}
\alias{pool}
\title{Pooling Function for PANIC (2010)}
\usage{
pool(p_values,test_values)
}
\arguments{
\item{p_values}{p_values supplied matrix containing p values}

\item{test_values}{The adf test to be pooled}
}
\value{
adf31a x P-value for the pooled test

adf31b x P-value for the pooled test
}
\description{
This function find the P values for the pooled test in PANIC (2010)
}

