% Generated by roxygen2 (4.0.2): do not edit by hand
\name{poolcoint}
\alias{poolcoint}
\title{Pooling Function for Cointegration test PANIC (2004)}
\usage{
poolcoint(a,x,r)
}
\arguments{
\item{a}{A supplied matrix containg the containing p values}

\item{x}{The adf test to be pooled}

\item{r}{The number of factors determined by getnfac()}
}
\value{
pvala a P-value for the pooled cointegration test

pvalb a P-value for the pooled cointegration test
}
\description{
This function find the P values for the pooled cointegration test in PANIC (2010)
}

