% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panic_generics.R
\name{summary.panic}
\alias{summary.panic}
\title{Summary function for panic objects}
\usage{
\method{summary}{panic}(object, ...)
}
\description{
Summarizes outputs and gives significance for the panic object
}

