% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimr.r
\name{trimr}
\alias{trimr}
\title{Trim dataset}
\usage{
trimr(x,a,b)
}
\arguments{
\item{x}{A NxT matrix containing the data to be trimmed}

\item{a}{number of columns to trim from matrix}

\item{b}{number of rows to trim from matrix}
}
\value{
xx the trimmed data set
}
\description{
This function trims the dataset by n and/or t rows and columns
}

