% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsd.r
\name{glsd}
\alias{glsd}
\title{General Least Squares Detrending}
\usage{
glsd(y,p)
}
\arguments{
\item{y}{A matrix containing the data to be detrended}

\item{p}{An integer with value of either 0 or 1 which decides the value of the penalty term, chat.
This is either either -7 or -13.5, respectively}
}
\value{
yt A matrix of the detrended data
}
\description{
This function detrends the data by general least squares.
}

