% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagn.R
\name{lagn}
\alias{lagn}
\title{Create lags for matrix}
\usage{
lagn(x,n)
}
\arguments{
\item{x}{A vector to apply the lag to}

\item{n}{The number of lags to add to the vector}
}
\value{
y A vector with a lag n added
}
\description{
This function adds lags to a vector.
}

