% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minindc.r
\name{minindc}
\alias{minindc}
\title{Create an Index of lowest values of each column}
\usage{
minindc(x)
}
\arguments{
\item{x}{A matrix that will be be used to create the index of lowest values.}
}
\value{
pos A vector containing an index of the row position of the lowest
value within each column of a matrix.
}
\description{
This function creates a vector filled with the row position
of the lowest value within each column of a matrix.
}

