% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydiff.r
\name{mydiff}
\alias{mydiff}
\title{Difference a matrix}
\usage{
mydiff(x,k)
}
\arguments{
\item{x}{The matrix to be differenced.}

\item{k}{An integer of the length of difference to perform on each vector of the NxT matrix.}
}
\value{
xx A matrix that has been differenced k times.
}
\description{
This function differences an entire matrix by some defined
constant k.
}

