% Generated by roxygen2 (4.0.2): do not edit by hand
\name{myols}
\alias{myols}
\title{Beta Coefficients for standard OLS}
\usage{
myols(x,y)
}
\arguments{
\item{x}{A NxT matrix of dependent variables.}

\item{y}{A 1xT vector that is used as the independent variable.}
}
\description{
Returns the Beta values for an Ordinary Least Squares.
}

