% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nuisance}
\alias{nuisance}
\title{Estimate the nuisance parameters of the error term}
\usage{
nuisance(res,k)
}
\arguments{
\item{res}{A matrix consisting of the residuals from a factor model.}

\item{k}{If fixk is 0, then automatic bandwidth selection is performed.
Otherwise, the integer placed here will be the selected bandwidth.}
}
\value{
sig2 The vector of short run variances

omega2 The vector of long run variances

half The vector of one-sided variances
}
\description{
This function estimates the short-run, long-run, and one sided
variance of the error term
}

