% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panic04}
\alias{panic04}
\title{PANIC (2004) Non-Stationarity Tests on Common and Idiosyncratic Components}
\usage{
panic04(x, nfac, k1, jj)
}
\arguments{
\item{x}{A NxT matrix containing the data}

\item{nfac}{An integer specifying the maximum number of factors allowed
while estimating the factor model.}

\item{k1}{The maximum lag allowed in the ADF test.}

\item{jj}{an Integer 1 through 8. Choices 1 through 7 are respectively, IC(1),
IC(2), IC(3), AIC(1), BIC(1), AIC(3), and BIC(3), respectively. Choosing 8
makes the number of factors equal to the number of columns whose sum of
eigenvalues is less than  or equal to .5.}
}
\value{
adff A data frame containing ADF p value, Demeaned Error term p value,Demeaned
 and detrended p value,R squared for principle component
 , And the significance of the Error components.

adf.ind A matrix containing the pooled Demeaned ADF test on the data, the
pooled ADF test on the common components, the pooled demeaned ADF test on the
Idiosyncratic component, and the pooled first differenced and demeand ADF test on the
Idiosyncratic component.
}
\description{
This function performs PANIC (2010) Model C, PAC, and PMSB tests.
PAC estimates the pooled autoregressive coefficient, PMSB uses a sample
moment, and Model C performs the MP test while projecting on intercept and trend.
 The sample moments test is based off of the modified Sargan-Bhargava test (PMSB).
}
\references{
Bai, Jushan, and Serena Ng.
"A PANIC Attack on Unit Roots and Cointegration."
Econometrica 72.4 (2004): 1127-1177. Print.
}

