% Generated by roxygen2 (4.0.2): do not edit by hand
\name{panic10}
\alias{panic10}
\title{PANIC (2010)}
\usage{
panic10(x, nfac, k1, jj, demean)
}
\arguments{
\item{x}{A NxT matrix containing the data}

\item{nfac}{An integer specifying the maximum number of factors allowed
while estimating the factor model.}

\item{k1}{The maximum lag allowed in the ADF test.}

\item{jj}{an Integer 1 through 8. Choices 1 through 7 are respectively, IC(1),
IC(2), IC(3), AIC(1), BIC(1), AIC(3), and BIC(3), respectively. Choosing 8
makes the number of factors equal to the number of columns whose sum of
eigenvalues is less than  or equal to .5.}

\item{demean}{logical argument. If TRUE, function performs tests on demeaned
data. If FALSE, uses non-demeanded data generating process.}
}
\value{
rho1 Estimation of the Pooled Autoregressive Coefficient.

Model This function shows MP models A, B, and C. A assumes no deterministic component.
B assumes a constant and allows for a fixed effect. C allows a fixed effect
and trend.

MP.tests If demean is false, a matrix containing results for models A and B.
If demean is true, returns results for Pa, Pb, and Model C

test.C.P A matrix containing the Pooled tests as well as Model C.

extra.test If demean is false, a matrix containing results PMSB test with fixed effects,
 rho, and the original ADF test on the idiosyncratic component from PANIC (2004).
If demean is true, returns results for PMSB test with fixed effects and trend, rho,
 and original LM test on idiosyncratic component from PANIC (2004)
}
\description{
This function performs the tests of PANIC (2010).
}
\references{
Bai, Jushan, and Serena Ng.
'Panel Unit Root Tests With Cross-Section Dependence: A Further Investigation.'
Econometric Theory 26.04 (2010): 1088-1114. Print.
}

