% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pc}
\alias{pc}
\title{Principle Component Analysis for PANIC (2004)}
\usage{
pc(y,nfac)
}
\arguments{
\item{y}{A NxT matrix containing the data}

\item{nfac}{An integer specifying the maximum number of factors allowed
while estimating the factor model.}
}
\value{
ehat The Idiosyncratic component of the factor model

fhat The approximate factors of the approximate factor model

lambda The factor loadings of the factor model
}
\description{
This function performs the principle component analysis in order to determine
the Common and Idiosyncratic Components of the factor model.
}

