% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pool}
\alias{pool}
\title{Pooling Function for PANIC (2010)}
\usage{
pool(a,x)
}
\arguments{
\item{a}{A supplied matrix containing p values}

\item{x}{The adf test to be pooled}
}
\value{
adf31a a P-value for the pooled test

adf31b a P-value for the pooled test
}
\description{
This function find the P values for the pooled test in PANIC (2010)
}

