% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool10.r
\name{pool}
\alias{pool}
\title{Pooling Function for PANIC (2010)}
\usage{
pool(p_values,test_values)
}
\arguments{
\item{p_values}{a data frame containing the p values}

\item{test_values}{a matrix of adf test values to be pooled}
}
\value{
adf31a a numeric vector of the fisher sum of the p-values

adf31b a numeric vector containing the critical value of the pooling test
}
\description{
This function find the P values for the pooled test in PANIC (2010)
}

