% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolcoint.R
\name{poolcoint}
\alias{poolcoint}
\title{Pooling Function for Cointegration test PANIC (2004)}
\usage{
poolcoint(a,x,r)
}
\arguments{
\item{a}{A matrix containing the p values}

\item{x}{A matrix containing the adf test to be pooled}

\item{r}{An integer for the number of factors determined by getnfac()}
}
\value{
pvala a numeric vector of the fisher sum of the p-values for the cointegration test

pvalb a numeric vector containing the critical value of the cointegration test
}
\description{
This function find the P values for the pooled cointegration test in PANIC (2010)
}

